refactor: extract navigation control
