+++ apps/client/src/common/components/navigation-menu/ViewNavigationMenu.tsx
@@ -15,72 +15,79 @@ import RenameClientModal from './rename-client-modal/RenameClientModal';
 import NavigationMenu from './NavigationMenu';
 
 import style from './NavigationMenu.module.scss';
+import FloatingNavigation from './FloatingNavigation';
 
 function ViewNavigationMenu() {
   const location = useLocation();
   const { fullscreen, toggle } = useFullscreen();
   const { toggleMirror } = useViewOptionsStore();
   const [searchParams, setSearchParams] = useSearchParams();
-  const { isOpen, onOpen, onClose } = useDisclosure();
+  const { isOpen: isRenameOpen, onOpen: onRenameOpen, onClose: onRenameClose } = useDisclosure();
+  const { isOpen: isMenuOpen, onOpen: onMenuOpen, onClose: onMenuClose } = useDisclosure();
 
   const showEditFormDrawer = useCallback(() => {
     searchParams.set('edit', 'true');
     setSearchParams(searchParams);
   }, [searchParams, setSearchParams]);
 
+  const toggleMenu = () => (isMenuOpen ? onMenuClose() : onMenuOpen());
+
   return (
-    <NavigationMenu editCallback={showEditFormDrawer}>
-      <RenameClientModal isOpen={isOpen} onClose={onClose} />
-      <div className={style.buttonsContainer}>
-        <div
-          className={style.link}
-          tabIndex={0}
-          role='button'
-          onClick={toggle}
-          onKeyDown={(event) => {
-            isKeyEnter(event) && toggle();
-          }}
-        >
-          Toggle Fullscreen
-          {fullscreen ? <IoContract /> : <IoExpand />}
+    <>
+      <FloatingNavigation toggleMenu={toggleMenu} toggleSettings={showEditFormDrawer} />
+      <NavigationMenu isOpen={isMenuOpen} onClose={onMenuClose}>
+        <RenameClientModal isOpen={isRenameOpen} onClose={onRenameClose} />
+        <div className={style.buttonsContainer}>
+          <div
+            className={style.link}
+            tabIndex={0}
+            role='button'
+            onClick={toggle}
+            onKeyDown={(event) => {
+              isKeyEnter(event) && toggle();
+            }}
+          >
+            Toggle Fullscreen
+            {fullscreen ? <IoContract /> : <IoExpand />}
+          </div>
+          <div
+            className={style.link}
+            tabIndex={0}
+            role='button'
+            onClick={() => toggleMirror()}
+            onKeyDown={(event) => {
+              isKeyEnter(event) && toggleMirror();
+            }}
+          >
+            Flip Screen
+            <IoSwapVertical />
+          </div>
+          <div
+            className={style.link}
+            tabIndex={0}
+            role='button'
+            onClick={onRenameOpen}
+            onKeyDown={(event) => {
+              isKeyEnter(event) && onRenameOpen();
+            }}
+          >
+            Rename Client
+          </div>
         </div>
-        <div
-          className={style.link}
-          tabIndex={0}
-          role='button'
-          onClick={() => toggleMirror()}
-          onKeyDown={(event) => {
-            isKeyEnter(event) && toggleMirror();
-          }}
-        >
-          Flip Screen
-          <IoSwapVertical />
-        </div>
-        <div
-          className={style.link}
-          tabIndex={0}
-          role='button'
-          onClick={onOpen}
-          onKeyDown={(event) => {
-            isKeyEnter(event) && onOpen();
-          }}
-        >
-          Rename Client
-        </div>
-      </div>
-      <hr className={style.separator} />
-      {navigatorConstants.map((route) => (
-        <Link
-          key={route.url}
-          to={route.url}
-          className={`${style.link} ${route.url === location.pathname ? style.current : undefined}`}
-          tabIndex={0}
-        >
-          {route.label}
-          <IoArrowUp className={style.linkIcon} />
-        </Link>
-      ))}
-    </NavigationMenu>
+        <hr className={style.separator} />
+        {navigatorConstants.map((route) => (
+          <Link
+            key={route.url}
+            to={route.url}
+            className={`${style.link} ${route.url === location.pathname ? style.current : undefined}`}
+            tabIndex={0}
+          >
+            {route.label}
+            <IoArrowUp className={style.linkIcon} />
+          </Link>
+        ))}
+      </NavigationMenu>
+    </>
   );
 }
